[jit] Improve the memory management.
Allocate both code and data to the same pages, which makes it easier to
manage the memory for AMD64 (and the upcoming ARM port). But make sure
to avoid placing data and code within the same 64-byte cache line (as
suggested by the AMD64 software optimization guide). This does of course
also work for processors with smaller L1 cache lines or no caches at all.